+++ services/sharepoint-connector/src/constants/sync-step.enum.ts
@@ -1,5 +1,11 @@
-export const SyncStep = {
+export const FullSyncStep = {
   SitesConfigLoading: 'sites_config_loading',
+  Unknown: 'unknown',
+} as const;
+
+export type FullSyncStep = (typeof FullSyncStep)[keyof typeof FullSyncStep];
+
+export const SiteSyncStep = {
   SiteNameFetch: 'site_name_fetch',
   RootScopeInit: 'root_scope_initialization',
   SubsiteDiscovery: 'subsite_discovery',
@@ -7,15 +13,14 @@ export const SyncStep = {
   SubsiteItemsFetch: 'subsite_items_fetch',
   ScopesCreation: 'scopes_creation',
   ContentSync: 'content_sync',
-  PermissionsSync: 'permissions_sync',
   PermissionsFetch: 'permissions_fetch',
   GroupsMembershipsFetch: 'groups_memberships_fetch',
   UniqueDataFetch: 'unique_data_fetch',
   GroupsSync: 'groups_sync',
   FilePermissionsSync: 'file_permissions_sync',
   FolderPermissionsSync: 'folder_permissions_sync',
+  UnknownPermissionsSync: 'unknown_permissions_sync',
   OrphanScopeCleanup: 'orphan_scope_cleanup',
-  Unknown: 'unknown',
 } as const;
 
-export type SyncStep = (typeof SyncStep)[keyof typeof SyncStep];
+export type SiteSyncStep = (typeof SiteSyncStep)[keyof typeof SiteSyncStep];